requirements.txt
torch==2.1.0
pyro-ppl==1.8.6
pandas>=2.0.0
numpy>=1.24.0
